refactor: replacing the name of handleSenderMessage
